refactor: reorganize imports in Sermon and EventPosts components for clarity
